+++ app/frontend/src/javascript/components/dashboard/reservations/prepaid-packs-panel.tsx
@@ -6,28 +6,45 @@ import { UserPack } from '../../../models/user-pack';
 import UserPackAPI from '../../../api/user-pack';
 import FormatLib from '../../../lib/format';
 import SettingAPI from '../../../api/setting';
+import { Machine } from '../../../models/machine';
+import MachineAPI from '../../../api/machine';
+import { SubmitHandler, useForm } from 'react-hook-form';
+import { FabButton } from '../../base/fab-button';
+import { FormSelect } from '../../form/form-select';
+import { SelectOption } from '../../../models/select';
+import { ProposePacksModal } from '../../prepaid-packs/propose-packs-modal';
+import * as React from 'react';
+import { User } from '../../../models/user';
 
 interface PrepaidPacksPanelProps {
-  userId: number,
+  user: User,
   onError: (message: string) => void
 }
 
 /**
  * List all available prepaid packs for the given user
  */
-const PrepaidPacksPanel: React.FC<PrepaidPacksPanelProps> = ({ userId, onError }) => {
+const PrepaidPacksPanel: React.FC<PrepaidPacksPanelProps> = ({ user, onError }) => {
   const { t } = useTranslation('logged');
 
+  const [machines, setMachines] = useState<Array<Machine>>([]);
   const [packs, setPacks] = useState<Array<UserPack>>([]);
   const [threshold, setThreshold] = useState<number>(null);
+  const [selectedMachine, setSelectedMachine] = useState<Machine>(null);
+  const [packsModal, setPacksModal] = useState<boolean>(false);
+
+  const { handleSubmit, control } = useForm<{ machine_id: number }>();
 
   useEffect(() => {
-    UserPackAPI.index({ user_id: userId })
-      .then(res => setPacks(res))
-      .catch(error => onError(error));
+    UserPackAPI.index({ user_id: user.id })
+      .then(setPacks)
+      .catch(onError);
     SettingAPI.get('renew_pack_threshold')
       .then(data => setThreshold(parseFloat(data.value)))
-      .catch(error => onError(error));
+      .catch(onError);
+    MachineAPI.index({ disabled: false })
+      .then(setMachines)
+      .catch(onError);
   }, []);
 
   /**
@@ -38,6 +55,41 @@ const PrepaidPacksPanel: React.FC<PrepaidPacksPanelProps> = ({ userId, onError }
       return pack.prepaid_pack.minutes - pack.minutes_used <= pack.prepaid_pack.minutes * threshold;
     }
     return pack.prepaid_pack.minutes - pack.minutes_used <= threshold * 60;
+  };
+
+  /**
+   * Callback triggered when the user clicks on "buy a pack"
+   */
+  const onBuyPack: SubmitHandler<{ machine_id: number }> = (data) => {
+    const machine = machines.find(m => m.id === data.machine_id);
+    setSelectedMachine(machine);
+    togglePacksModal();
+  };
+
+  /**
+   * Open/closes the buy pack modal
+   */
+  const togglePacksModal = () => {
+    setPacksModal(!packsModal);
+  };
+
+  /**
+   * Build the options for the select dropdown, for the given list of machines
+   */
+  const buildMachinesOptions = (machines: Array<Machine>): Array<SelectOption<number>> => {
+    return machines.map(m => {
+      return { label: m.name, value: m.id };
+    });
+  };
+
+  /**
+   * Callback triggered when a prepaid pack was successfully bought: refresh the list of packs for the user
+   */
+  const onPackBoughtSuccess = () => {
+    togglePacksModal();
+    UserPackAPI.index({ user_id: user.id })
+      .then(setPacks)
+      .catch(onError);
   };
 
   return (
@@ -57,7 +109,7 @@ const PrepaidPacksPanel: React.FC<PrepaidPacksPanelProps> = ({ userId, onError }
               <p className="countdown"><span>{pack.minutes_used / 60}H</span> / {pack.prepaid_pack.minutes / 60}H</p>
             </div>
           </div>
-
+          { /* usage history is not saved for now
           <div className="prepaid-packs-list is-history">
             <span className='prepaid-packs-list-label'>{t('app.logged.dashboard.reservations_dashboard.prepaid_packs_panel.history')}</span>
 
@@ -66,12 +118,28 @@ const PrepaidPacksPanel: React.FC<PrepaidPacksPanelProps> = ({ userId, onError }
               <p className="date">00/00/00</p>
             </div>
           </div>
+          */ }
         </div>
       ))}
 
       <div className='prepaid-packs-cta'>
         <p>{t('app.logged.dashboard.reservations_dashboard.prepaid_packs_panel.cta_info')}</p>
-        <button className='fab-button is-black'>{t('app.logged.dashboard.reservations_dashboard.prepaid_packs_panel.cta_button')}</button>
+        <form onSubmit={handleSubmit(onBuyPack)}>
+          <FormSelect options={buildMachinesOptions(machines)} control={control} id="machine_id" rules={{ required: true }} label={t('app.logged.dashboard.reservations_dashboard.prepaid_packs_panel.select_machine')} />
+          <FabButton className='is-black' type="submit">
+            {t('app.logged.dashboard.reservations_dashboard.prepaid_packs_panel.cta_button')}
+          </FabButton>
+        </form>
+        {selectedMachine && packsModal &&
+          <ProposePacksModal isOpen={packsModal}
+                             toggleModal={togglePacksModal}
+                             item={selectedMachine}
+                             itemType='Machine'
+                             customer={user}
+                             operator={user}
+                             onError={onError}
+                             onDecline={togglePacksModal}
+                             onSuccess={onPackBoughtSuccess} />}
       </div>
 
     </FabPanel>