+++ app/frontend/src/javascript/components/dashboard/reservations/reservations-dashboard.tsx
@@ -8,18 +8,19 @@ import { SettingName } from '../../../models/setting';
 import { CreditsPanel } from './credits-panel';
 import { useTranslation } from 'react-i18next';
 import { PrepaidPacksPanel } from './prepaid-packs-panel';
+import { User } from '../../../models/user';
 
 declare const Application: IApplication;
 
 interface ReservationsDashboardProps {
   onError: (message: string) => void,
-  userId: number
+  user: User
 }
 
 /**
  * User dashboard showing everything about his spaces/machine reservations and also remaining credits
  */
-const ReservationsDashboard: React.FC<ReservationsDashboardProps> = ({ onError, userId }) => {
+const ReservationsDashboard: React.FC<ReservationsDashboardProps> = ({ onError, user }) => {
   const { t } = useTranslation('logged');
   const [modules, setModules] = useState<Map<SettingName, string>>();
 
@@ -33,17 +34,17 @@ const ReservationsDashboard: React.FC<ReservationsDashboardProps> = ({ onError, 
     <div className="reservations-dashboard">
       {modules?.get('machines_module') !== 'false' && <div className="section">
         <p className="section-title">{t('app.logged.dashboard.reservations_dashboard.machine_section_title')}</p>
-        <CreditsPanel userId={userId} onError={onError} reservableType="Machine" />
-        <PrepaidPacksPanel userId={userId} onError={onError} />
-        <ReservationsPanel userId={userId} onError={onError} reservableType="Machine" />
+        <CreditsPanel userId={user.id} onError={onError} reservableType="Machine" />
+        <PrepaidPacksPanel user={user} onError={onError} />
+        <ReservationsPanel userId={user.id} onError={onError} reservableType="Machine" />
       </div>}
       {modules?.get('spaces_module') !== 'false' && <div className="section">
         <p className="section-title">{t('app.logged.dashboard.reservations_dashboard.space_section_title')}</p>
-        <CreditsPanel userId={userId} onError={onError} reservableType="Space" />
-        <ReservationsPanel userId={userId} onError={onError} reservableType="Space" />
+        <CreditsPanel userId={user.id} onError={onError} reservableType="Space" />
+        <ReservationsPanel userId={user.id} onError={onError} reservableType="Space" />
       </div>}
     </div>
   );
 };
 
-Application.Components.component('reservationsDashboard', react2angular(ReservationsDashboard, ['onError', 'userId']));
+Application.Components.component('reservationsDashboard', react2angular(ReservationsDashboard, ['onError', 'user']));